Improve time complexity of iterating through claims
Before the change, the time complexity was O(n^2) because the code iterated over all the principal claims, and for each claim, it iterated over all the excluded claims.
After the change, the time complexity has become O(n) because the code iterates over all the principal claims, and for each claim, it accesses a hashed item of the excluded claims which is O(1).